+++ backend/build.gradle
@@ -51,6 +51,10 @@ dependencies {
 
     // jackson 관련
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.12.3'
+
+    // google guava
+    implementation group: 'com.google.guava', name: 'guava', version: '30.1.1-jre'
+
 }
 
 test {
