+++ backend/src/test/java/com/darass/darass/AcceptanceTest.java
@@ -1,73 +1,42 @@
 package com.darass.darass;
 
-import com.darass.darass.comment.repository.CommentRepository;
-import com.darass.darass.project.repository.ProjectRepository;
-import com.darass.darass.user.repository.UserRepository;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.junit.jupiter.api.parallel.Isolated;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.restdocs.mockmvc.RestDocumentationResultHandler;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.filter.CharacterEncodingFilter;
 
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.documentationConfiguration;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-
+@Isolated
 @ExtendWith({RestDocumentationExtension.class, SpringExtension.class})
-@SpringBootTest
-@ActiveProfiles("test")
-public class AcceptanceTest {
+public class AcceptanceTest extends SpringContainerTest {
 
     protected MockMvc mockMvc;
-
-    @Autowired
-    private UserRepository users;
-    @Autowired
-    private ProjectRepository projects;
-    @Autowired
-    private CommentRepository comments;
-
-    private RestDocumentationResultHandler document;
 
     protected static String asJsonString(Object obj) throws JsonProcessingException {
         return new ObjectMapper().writeValueAsString(obj);
     }
 
-    protected <T> T asObject(String json, Class<T> objectClass) throws JsonProcessingException {
-        return new ObjectMapper().readValue(json, objectClass);
-    }
-
     @BeforeEach
     public void setUp(WebApplicationContext webApplicationContext,
-                      RestDocumentationContextProvider restDocumentation) {
+        RestDocumentationContextProvider restDocumentation) throws Exception {
 
         this.mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext)
-                .addFilters(new CharacterEncodingFilter("UTF-8", true))
-                .apply(documentationConfiguration(restDocumentation)
-                    .operationPreprocessors()
-                    .withRequestDefaults(prettyPrint())
-                    .withResponseDefaults(prettyPrint()))
-                .alwaysDo(MockMvcResultHandlers.print())
-                .build();
+            .addFilters(new CharacterEncodingFilter("UTF-8", true))
+            .apply(documentationConfiguration(restDocumentation)
+                .operationPreprocessors()
+                .withRequestDefaults(prettyPrint())
+                .withResponseDefaults(prettyPrint()))
+            .build();
     }
 
-    @AfterEach
-    public void tearDown() {
-        comments.deleteAll();
-        projects.deleteAll();
-        users.deleteAll();
-    }
 }
