+++ backend/src/test/java/com/darass/darass/comment/domain/CommentsTest.java
@@ -2,8 +2,8 @@ package com.darass.darass.comment.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.darass.darass.project.domain.CustomSecretKeyFactory;
 import com.darass.darass.project.domain.Project;
-import com.darass.darass.project.service.CustomSecretKeyFactory;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
