+++ backend/src/test/java/com/darass/darass/user/service/UserServiceTest.java
@@ -2,6 +2,7 @@ package com.darass.darass.user.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.darass.darass.SpringContainerTest;
 import com.darass.darass.comment.controller.dto.UserResponse;
 import com.darass.darass.exception.ExceptionWithMessageAndCode;
 import com.darass.darass.user.controller.dto.UserUpdateRequest;
@@ -13,13 +14,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 @DisplayName("UserService 클래스")
-@ActiveProfiles("test")
-@SpringBootTest
-class UserServiceTest {
+class UserServiceTest extends SpringContainerTest {
 
     @Autowired
     private UserService userService;
