+++ backend/src/test/resources/application.properties
@@ -1,0 +1,6 @@
+spring.main.banner-mode=off
+spring.jpa.show-sql=false
+spring.jpa.generate-ddl=false
+logging.level.root=error
+logging.level.org.springframework=error
+logging.level.org.hibernate.SQL= error
